+++ db/schema.rb
@@ -296,6 +296,7 @@ ActiveRecord::Schema[7.1].define(version: 2026_07_09_130000) do
     t.index ["home_team_id"], name: "index_games_on_home_team_id"
     t.index ["legacy_ref"], name: "index_games_on_legacy_ref", unique: true, where: "(legacy_ref IS NOT NULL)"
     t.index ["officiating_referee_ids"], name: "index_games_on_officiating_referee_ids", using: :gin
+    t.index ["players"], name: "index_games_on_players", using: :gin
     t.index ["referee_ids"], name: "index_games_on_referee_ids", using: :gin
   end